+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add logging to unified file (gh-324).
 - Add memory leak detection during server process execution (gh-349).
+- Improve `luatest.log` function if a `nil` value is passed (gh-360).
 
 ## 1.0.1